requirements.txt
streamlit
crewai[google-genai]
python-dotenv